feat: Remove agent_script data source to simplify resources (#8)
This adds a new resource for associating instance IDs, which will only
need to be used in edge-cases. We'll auto-detect instance identifiers
for the most popular cases, and perform zero-trust authentication.

It seemed weird to separate the agent script and agent itself. This also
led to difficulties previewing the operating system prior to start.