feat: Add app support
This adds what are presently named "devurls" in v1.
It seems this is a dated term, since this allows
much more than accessing applications via URL.

"coder open <name>" will launch any apps defined.
If in the web, it'll open either a web terminal or
port forward to the desired application. If in the terminal,
it'll open the browser, or launch the command over SSH.