fix: Add custom host for access URL and remove access port
This allows a host override, which is useful for running Docker
containers that need to access the same deployment port, but on
`host.docker.internal` as the hostname.

It also removes access port, which was released a few days ago in hopes of
solving the same problem, but never really did. It doesn't seem likely to
be used yet, so it feels safe to remove.