feat: add `enabled` computed field to `coder_ai_task`
As discussed, we want to enable a way for consumers to know if their
template is being provisioned as a task or not. We know that
`CODER_TASK_ID` is set by the provisioner, it is being provisioned as
a task, and not as a usual workspace. We use this knowledge to set
a computed field `enabled` to reflect this.

fix: tests

fix: maybe fix?

does this fix it?