feat: add `enabled` computed field to `coder_ai_task` (#451)
As discussed, we want to enable a way for consumers to know if their
template is being provisioned as a task or not. We know that if 
`CODER_TASK_ID` is set by the provisioner, it is being provisioned as
a task, and not as workspace. We use this knowledge to set a computed 
field `enabled` to reflect this.